Addressing comments, and refactoring logging code
Renames neurobooth.logging as neurobooth.log_manager to avoid interpreter and human confusion with python logging.

Also,
- introduces argparse to handle parsing for transfer_data.
- introduces a types module to handle common defintions
