+++ neurobooth_os/netcomm/__init__.py
@@ -4,5 +4,6 @@ from .server import (
     get_fprint,
     get_messages_to_ctr,
     NewStdout,
-    get_data_timeout,
-)
+    get_data_timeout)
+from .types import NODE_NAMES
+
