+++ neurobooth_os/netcomm/types.py
@@ -1,0 +1,6 @@
+""""
+    Definitions of commonly used types and constants
+"""
+
+#  System nodes. Each node represents a machine in a single Neurobooth environment
+NODE_NAMES = ['control', 'presentation', 'acquisition']
